refactor: unify footer helpers
