+++ apps/web/components/blocks/Footer/types.ts
@@ -1,3 +1,5 @@
+import type { Block } from '@plentymarkets/shop-api';
+
 export type FooterProps = {
   content?: FooterSettings;
 };
@@ -28,3 +30,11 @@ export interface FooterSettings {
   footnoteAlign: 'left' | 'center' | 'right';
   colors: FooterSettingsColors;
 }
+
+type AddFooterBlockOptions = {
+  data: Ref<Block[]>;
+  cachedFooter: Ref<unknown>;
+  cleanData?: Ref<Block[]>;
+};
+
+export type AddFooterBlock = (options: AddFooterBlockOptions) => void;
